some fixes, spok still broken

+++ src/keys/cl03_key.rs
@@ -41,7 +41,7 @@ impl CL03CommitmentPublicKey {
 
     //verifier_pk -> N = issuer_pk.N
     //trusted_party_pk -> N = None
-    pub fn generate<CS: CLCiphersuite>(N: Option<Integer>, n_attributes: Option<u32>) -> Self{
+    pub fn generate<CS: CLCiphersuite>(N: Option<Integer>, n_attributes: Option<usize>) -> Self{
         let n = CS::SECPARAM; //SECPARAM
         let n_attributes = n_attributes.unwrap_or(1);
         let N = N.unwrap_or_else(|| {
